requirements.txt
typer
xgboost
pandas
statsmodels
scikit-learn
scipy
numpy
requests
pyyaml
git+https://github.com/RedLicorice/pyti.git
sqlalchemy
confuse
python-dotenv